Allow to mark accomplishment status

+++ db/migrate/20180220095521_create_learning_topics.rb
@@ -4,7 +4,7 @@ class CreateLearningTopics < ActiveRecord::Migration[5.1]
       t.string :title
       t.integer :course_id
       t.integer :parent_topic_id
-      t.float :position
+      t.integer :position
 
       t.timestamps
     end
